@@ -21,50 +21,46 @@
     "serve:ssr:demo": "node dist/demo/server/server.mjs"
   },
   "dependencies": {
-    "@angular/animations": "^17.3.0",
-    "@angular/common": "^17.3.0",
-    "@angular/compiler": "^17.3.0",
-    "@angular/core": "^17.3.0",
-    "@angular/forms": "^17.3.0",
-    "@angular/platform-browser": "^17.3.0",
-    "@angular/platform-browser-dynamic": "^17.3.0",
-    "@angular/platform-server": "^17.3.0",
-    "@angular/router": "^17.3.0",
-    "@angular/ssr": "^17.3.7",
-    "express": "^4.18.2",
+    "@angular/animations": "^20.0.0",
+    "@angular/common": "^20.0.0",
+    "@angular/compiler": "^20.0.0",
+    "@angular/core": "^20.0.0",
+    "@angular/forms": "^20.0.0",
+    "@angular/platform-browser": "^20.0.0",
+    "@angular/platform-browser-dynamic": "^20.0.0",
+    "@angular/platform-server": "^20.0.0",
+    "@angular/router": "^20.0.0",
+    "@angular/ssr": "^20.0.5",
+    "express": "^5.1.0",
     "rxjs": "~7.8.0",
     "tslib": "^2.3.0",
-    "zone.js": "~0.14.3"
+    "zone.js": "~0.15.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.7",
-    "@angular-eslint/builder": "^17.4.0",
-    "@angular-eslint/eslint-plugin": "^17.4.0",
-    "@angular-eslint/eslint-plugin-template": "^17.4.0",
-    "@angular-eslint/schematics": "^17.4.0",
-    "@angular-eslint/template-parser": "^17.4.0",
-    "@angular/cli": "^17.3.7",
-    "@angular/compiler-cli": "^17.3.0",
+    "@angular-devkit/build-angular": "^20.0.5",
+    "@angular/cli": "^20.0.5",
+    "@angular/compiler-cli": "^20.0.0",
     "@commitlint/cli": "^19.8.1",
     "@commitlint/config-conventional": "^19.8.1",
-    "@types/express": "^4.17.23",
+    "@types/express": "^5.0.1",
     "@types/jasmine": "~5.1.0",
-    "@types/node": "^22.16.0",
+    "@types/node": "^20.17.19",
+    "angular-eslint": "^20.1.1",
     "bootstrap": "^5.3.7",
-    "eslint": "^8.57.1",
-    "eslint-config-pegasus": "^5.0.0",
+    "eslint": "^9.30.1",
+    "eslint-config-pegasus": "^6.0.2",
     "husky": "^9.1.7",
     "is-ci": "^4.1.0",
-    "jasmine-core": "~5.1.0",
+    "jasmine-core": "~5.7.0",
     "karma": "~6.4.0",
     "karma-chrome-launcher": "~3.2.0",
     "karma-coverage": "~2.2.0",
     "karma-jasmine": "~5.1.0",
     "karma-jasmine-html-reporter": "^2.1.0",
     "lint-staged": "^16.1.2",
-    "ng-packagr": "^17.3.0",
+    "ng-packagr": "^20.0.0",
     "picocolors": "^1.1.1",
     "prettier": "^3.6.2",
-    "typescript": "~5.4.2"
+    "typescript": "~5.8.2"
   }
 }
